+++ olivia/src/services/ContextManagerService.py
@@ -1,9 +1,11 @@
+from src.aux.utils import open_file
+
 class ContextManagerService:
   def __init__(self, language):
     self.language = language
     self.context = None
   
-  def get_context_file(self):
+  def get_context(self):
 
     if self.language == "en":
         self.context = "chatbot_en.txt"
@@ -13,5 +15,4 @@ class ContextManagerService:
       # TODO: create custom exceptions
         raise ValueError(f"Invalid Context. We currently support en (english) and pt (brazililian portuguese), but got {LANGUAGE} language")
 
-    return self.context
-
+    return open_file(self.context)
